DPM 테스트 소수점도 계산할 수 있도록 수정 및 테스트몹 저항 1로 변경

+++ src/main/java/com/elland/dev/ellandtest/EllandTest.java
@@ -9,7 +9,7 @@ import org.bukkit.plugin.java.JavaPlugin;
 import java.util.HashMap;
 
 public final class EllandTest extends JavaPlugin {
-    public static HashMap<String, Integer> dpmMap = new HashMap<>();
+    public static HashMap<String, Double> dpmMap = new HashMap<>();
 
     @Override
     public void onEnable() {
